first

+++ server/index.ts
@@ -7,29 +7,11 @@ const app = express();
 app.use( cors());
 app.use(express.json())
 
-function main(){
-	// // Routing
-	// Routes setup
-	// const users = require('./routes/users.ts'); app.use('/api/users', users);
-	const test = require('./routes/test.ts'); app.use('/api/test', test);
-
-	const port = process.env.PORT || 4200;
-
-	console.log("darkness");
-	app.listen(port, () => console.log(`Server started at ${port}`));
-}
 
 // DB setup
-/* In the event the tables/database is not setup/wrong size, this should take care of that */
-/* Note: this means that if the models get change, then the program will delete the table*/
-/* And replace it with the new model */
-/* TODO once we get this in production, we have to disable this behavior. Otherwise*/
-/* We will be made an example of for crap developers */
-/* So be wary of this section in production: please please please please please*/
-/* remember this, and don't gloss over it */
-
-// DB Models
 const Test = require('./model/Test.ts');
+
+// Dtabase initalize
 async function initialize() {
 	// create db if it doesn't already exist
 	const db = require('./config/keys.ts');
@@ -50,18 +32,24 @@ async function initialize() {
 		})
 		// Creating tables based on models
 
+		/* In the event the tables/database is not setup/ or change size, this overrides that */
+		/* Note: this means that if the models get change, then the program will delete the table*/
+		/* And replace it with the new model */
+		/* TODO once we get this in production, we have to disable this behavior. Otherwise*/
+		/* We will be made an example of for crap developers */
+		/* So be wary of this section in production: please please please please please*/
+		/* remember this, and don't gloss over it */
 
 		// sequelize.sync()
-		sequelize.sync({force:true})
-		.then(() => {
-			// Model synicng
-			if(Test == sequelize.models.Project)
-				console.log("NICE");
-			else{
-				console.log("The table is not equal to the sql represntation: migrate?")
-			}
-			Test.sync({force:true});
-		});
+		if( process.env.DEV == "true"){
+			sequelize.sync()
+		}
+		else{
+			sequelize.sync({force:true})
+			.then(() => {
+				Test.sync({force:true});
+			});
+		}
 	})
 	.catch(err =>{
 		console.log(err);
@@ -75,4 +63,13 @@ async function initialize() {
 }
 
 initialize();
+// // Routing
+// Routes setup
+// const users = require('./routes/users.ts'); app.use('/api/users', users);
+const test = require('./routes/test.ts'); app.use('/api/test', test);
+
+const port = process.env.PORT || 4200;
+
+console.log("darkness");
+app.listen(port, () => console.log(`Server started at ${port}`));
 export {}
